added field property 'db_decimals'
This allows you to specify the number of decimals used in the database,
in case it is different from the use in the model. For example, if the
data_type specifies "decimal:2" but the database defines "decimal:10,3",
you can set "db_decimals" to 3. This is important for performance
reasons, currently, the ORM will flag these fields as "changed", even if
they are not (only the number of decimals differ!)