don't do loose type comparisons on update
If there are type differences between stored data and used data, the
Observer_Typing should be used to make sure in both cases the data types
are correct.

In the case of update(), observers have run when column data is added to
the query, therefore the type should always be correct.

In the case of is_changed(), do a loose comparison on int and bool if
observers haven't run first (the default), to limit the number of false
results due to type mismatches.