do not overwrite cached objects on subsequent queries
Fixed a bug where related objects in cache were overwritten with newy
queried data. This not only caused altered/runtime data to be reset, but
also possibly incorrect, as observers where not triggered.
For objects that are already in cache, only update object properties not
in cache yet (for example due to an earlier partial select).