bugfix: ORM relations aren't "set" until they're loaded.
if you relied on `isset($model->relation)` or `empty($model->relation)`
in your code, and the relation was included in the query, it would fire
an additional lazy load query if model object that didn't have a related
record (which would always return zero rows).

Note that this fix is not backwards compatible if you didn't include the
related object in the query, and relied on this mechanism to trigger the
lazy load.

Note also that is soon as you access the `$model->relation` variable, it
will trigger a lazy load, so the impact of this change should be very
minimal, but would potentially save a lot of database I/O.