fix relation lazy load BC break
This is related to 588126c618f4f889caf014c852a77f1ddb780e7c

As that commit says, it will break BC for applications that expect lazy
load to work, or where the dev has been lazy and relied on the lazy load
feature to get related objects if and when needed, without explicitly
fetching them.

This comment makes the behaviour configurable, so you can revert back to
the old behaviour when you update the framework in an existing
application.